full admin  done

+++ controllers/admin.py
@@ -1,5 +1,6 @@
 from flask import jsonify, redirect, url_for
 from utils import db
+import traceback
 
 def admin_delete_users_action(request, session):
     try:
@@ -14,9 +15,19 @@ def admin_delete_users_action(request, session):
             delete from `student_subject` where user_id = {id}
         ''')
 
+        db.execute_query(f'''
+            delete from `marks`
+            where student_id = {id}
+        ''')
+
+        db.execute_query(f'''
+            delete from `attendance`
+            where student_id = {id}
+        ''')
+
         return redirect(url_for(f'admin_{role}s', success=f"{role} deleted successfully"))
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         return redirect(url_for(f'admin_{role}s', error="some error occured"))
 
 def admin_attendance_action(request, session):
@@ -44,8 +55,26 @@ def admin_attendance_action(request, session):
 
         return redirect(url_for(f'admin_attendance', success=f"attendance marked successfully"))
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         return redirect(url_for('admin_students', error="some error occured"))
+
+def admin_marks_action(request, session):
+    try:
+        student_id = request.args.get('student_id', "")
+        subject_id = request.args.get('subject_id', "")
+        marks = request.args.get('marks', "")
+
+        db.execute_query(f'''
+            update `marks`
+            set marks = {marks}
+            where student_id = {student_id}
+            and subject_id = {subject_id}
+        ''')
+
+        return redirect(url_for(f'admin_marks', success="Marks updated successfully"))
+    except:
+        print(traceback.print_exc())
+        return redirect(url_for(f'admin_marks', error="some error occured"))
 
 def admin_add_student_subject_action(request, session):
     try:
@@ -57,9 +86,14 @@ def admin_add_student_subject_action(request, session):
             values({subject_id}, {user_id})
         ''')
 
+        db.execute_query(f'''
+            insert into `marks`(subject_id, student_id, marks)
+            values({subject_id}, {user_id}, 0)
+        ''')
+
         return redirect(url_for(f'admin_edit_users', id=user_id, success=f"subject added successfully"))
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         return redirect(url_for(f'admin_edit_users', id=user_id, error="some error occured"))
 
 def admin_delete_student_subject_action(request, session):
@@ -67,12 +101,34 @@ def admin_delete_student_subject_action(request, session):
         id = request.args.get('id', "")
         ss_id = request.args.get('ss_id', "")
 
-        db.execute_query(f'''
-            delete from `student_subject` where id = {ss_id}
+        dt = db.execute_query(f'''
+            select user_id, subject_id from `student_subject` where id = {ss_id}
         ''')
 
+        print(dt)
+
+        if(dt):
+
+            dt = dt[0]
+
+            db.execute_query(f'''
+                delete from `student_subject` where id = {ss_id}
+            ''')
+
+            db.execute_query(f'''
+                delete from `marks`
+                where subject_id = {dt[1]}
+                and student_id = {dt[0]}
+            ''')
+
+            db.execute_query(f'''
+                delete from `attendance`
+                where subject_id = {dt[1]}
+            ''')
+
         return redirect(url_for('admin_edit_users', id=id, success="Subject deleted successfully"))
-    except:
+    except Exception as e:
+        print(traceback.print_exc())
         return redirect(url_for('admin_edit_users', id=id, error="some error occured"))
 
 def admin_create_subject_action(request, session):
@@ -87,7 +143,7 @@ def admin_create_subject_action(request, session):
 
         return redirect(url_for('admin_subjects', success="Subject created successfully"))
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         return redirect(url_for('admin_subjects', error="some error occured"))
 
 def admin_edit_subject_action(request, session):
@@ -107,7 +163,7 @@ def admin_edit_subject_action(request, session):
 
         return redirect(url_for('admin_subjects', success="Subject edited successfully"))
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         return redirect(url_for('admin_subjects', error="some error occured"))
 
 def admin_create_user_action(request, session):
@@ -122,7 +178,7 @@ def admin_create_user_action(request, session):
 
         return redirect(url_for(f'admin_{role}s', success=f"{role} created successfully"))
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         return redirect(url_for(f'admin_{role}s', error="some error occured"))
 
 def admin_edit_users_action(request, session):
@@ -142,7 +198,7 @@ def admin_edit_users_action(request, session):
 
         return redirect(url_for(f'admin_{role}s', success=f"{role} edited successfully"))
     except Exception as e:
-        print(e)
+        print(traceback.print_exc())
         return redirect(url_for(f'admin_{role}s', error="some error occured"))
 
 def admin_delete_subject_action(request, session):
@@ -154,6 +210,16 @@ def admin_delete_subject_action(request, session):
         
         db.execute_query(f'''
             delete from `student_subject` where subject_id = {subject}
+        ''')
+
+        db.execute_query(f'''
+            delete from `marks`
+            where subject_id = {subject}
+        ''')
+
+        db.execute_query(f'''
+            delete from `attendance`
+            where subject_id = {subject}
         ''')
 
         return redirect(url_for('admin_subjects', success="Subject deleted successfully"))
